cms/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
pydantic==2.5.0
python-dateutil==2.8.2
Pillow==10.1.0
bcrypt==4.1.1
filelock==3.13.1

